Authentication & authorization

+++ products/views.py
@@ -14,34 +14,45 @@ def product_details(request, pk):
 
 
 def product_add(request):
-    if request.method == 'POST':
-        form = AddProductForm(request.POST, request.FILES)
-        if form.is_valid():
-            form.save()
-            return render(request, 'products/product-add-successful.html')
+    if request.user.is_authenticated and request.user.is_superuser:
+
+        if request.method == 'POST':
+            form = AddProductForm(request.POST, request.FILES)
+            if form.is_valid():
+                form.save()
+                return render(request, 'products/product-add-successful.html')
+        else:
+
+            form = AddProductForm()
+
+        return render(request, 'products/product-add.html', {'form': form})
     else:
-
-        form = AddProductForm()
-
-    return render(request, 'products/product-add.html', {'form': form})
+        return redirect('product-list')
 
 
 def product_edit(request, pk):
-    product = get_object_or_404(Product, pk=pk)
+    if request.user.is_authenticated and request.user.is_superuser:
+        product = get_object_or_404(Product, pk=pk)
+        if request.method == 'POST':
+            form = AddProductForm(
+                request.POST, request.FILES,  instance=product)
+            if form.is_valid():
+                form.save()
+                return render(request, 'products/product-add-successful.html')
+        else:
 
-    if request.method == 'POST':
-        form = AddProductForm(request.POST, request.FILES,  instance=product)
-        if form.is_valid():
-            form.save()
-            return render(request, 'products/product-add-successful.html')
+            form = AddProductForm(instance=product)
+
+        return render(request, 'products/product-add.html', {'form': form})
     else:
-
-        form = AddProductForm(instance=product)
-
-    return render(request, 'products/product-add.html', {'form': form})
+        return redirect('product-list')
 
 
 def product_delete(request, pk):
-    product = get_object_or_404(Product, pk=pk)
-    product.delete()
-    return redirect('product-list')
+    if request.user.is_authenticated and request.user.is_superuser:
+
+        product = get_object_or_404(Product, pk=pk)
+        product.delete()
+        return redirect('product-list')
+    else:
+        return redirect('product-list')
